Proxy temporary and permanent dashboards to ES
Allow temporary and permanent dashboards to be saved and loaded from
Elasticsearch. These routes are a combination of my own testing and this
example:

https://github.com/elasticsearch/kibana/blob/v3.1.0/sample/nginx.conf#L46-63

The index name is validated against the `Kibana.kibana_index` name to
prevent access to anything other than Kibana's internal index. This will
continue to work when changed from the default of `kibana-int`. As
demonstrated by the additional tests.