Replace params[:captures] with path_info
This gives us the same value - which is the path of the route always without
the `/kibana` prefix of where the rack app is mounted to - without relying
on regexes for routes.

It's worth noting that this isn't tested by the specs which start the rack
app on `/`. I'm not sure how it could be tested.